Closures work => Y works!
Compiled lambda were only keeping the list of frames that it was created
in, not the list of frames *and* the store. When a lambda was evaluated
and bound arguments to parameters, the store from the enclosing scope
(i.e. lambda) wasn't passed to it. Naturally, bindings from that
enclosing scope were missing. So, before you test for Y, test for
closures.